Differenciating CApp Artifacts

+++ components/mediation-admin/org.wso2.carbon.endpoint/src/main/java/org/wso2/carbon/endpoint/service/EndpointAdmin.java
@@ -27,11 +27,13 @@ import org.wso2.carbon.mediation.initializer.AbstractServiceBusAdmin;
 import org.wso2.carbon.mediation.initializer.ServiceBusConstants;
 import org.wso2.carbon.mediation.initializer.ServiceBusUtils;
 import org.wso2.carbon.mediation.initializer.persistence.MediationPersistenceManager;
+import org.wso2.carbon.mediation.initializer.services.CAppArtifactDataService;
 import org.wso2.carbon.registry.core.Resource;
 import org.wso2.carbon.registry.core.exceptions.RegistryException;
 import javax.xml.namespace.QName;
 import javax.xml.stream.XMLStreamException;
 import java.io.ByteArrayOutputStream;
+import java.io.File;
 import java.lang.String;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -78,7 +80,11 @@ public class EndpointAdmin extends AbstractServiceBusAdmin {
             endpointName = endpointName.trim();
             Endpoint ep = getSynapseConfiguration().getEndpoint(endpointName);
             ep.getContext().switchOn();
-            persistEndpoint(ep);
+            CAppArtifactDataService cAppArtifactDataService = ConfigHolder.getInstance().
+                    getcAppArtifactDataService();
+            if (!cAppArtifactDataService.isArtifactDeployedFromCApp(getTenantId(), ServiceBusConstants.ENDPOINT_TYPE + File.separator + ep.getName())) {
+                persistEndpoint(ep);
+            }
 
             if (log.isDebugEnabled()) {
                 log.debug("Endpoint " + ep.getName() + " switched on");
@@ -105,7 +111,11 @@ public class EndpointAdmin extends AbstractServiceBusAdmin {
             endpointName = endpointName.trim();
             Endpoint ep = getSynapseConfiguration().getEndpoint(endpointName);
             ep.getContext().switchOff();
-            persistEndpoint(ep);
+            CAppArtifactDataService cAppArtifactDataService = ConfigHolder.getInstance().
+                    getcAppArtifactDataService();
+            if (!cAppArtifactDataService.isArtifactDeployedFromCApp(getTenantId(), ServiceBusConstants.ENDPOINT_TYPE + File.separator + endpointName)) {
+                persistEndpoint(ep);
+            }
 
             if (log.isDebugEnabled()) {
                 log.debug("Endpoint " + ep.getName() + " switched off");
@@ -225,23 +235,29 @@ public class EndpointAdmin extends AbstractServiceBusAdmin {
         final Lock lock = getLock();
         try {
             lock.lock();
+            CAppArtifactDataService cAppArtifactDataService = ConfigHolder.getInstance().
+                    getcAppArtifactDataService();
             for(String endpointName : endpointNames){
                 assertNameNotEmpty(endpointName);
                 endpointName = endpointName.trim();
-                if (log.isDebugEnabled()) {
-                    log.debug("Deleting endpoint : " + endpointName + " from the configuration");
-                }
-                SynapseConfiguration synapseConfiguration = getSynapseConfiguration();
-                Endpoint endpoint = synapseConfiguration.getDefinedEndpoints().get(endpointName);
-                synapseConfiguration.removeEndpoint(endpointName);
-                MediationPersistenceManager pm = getMediationPersistenceManager();
-                String fileName = null;
-                if (endpoint instanceof AbstractEndpoint) {
-                    fileName = endpoint.getFileName();
-                }
-                pm.deleteItem(endpointName, fileName, ServiceBusConstants.ITEM_TYPE_ENDPOINT);
-                if (log.isDebugEnabled()) {
-                    log.debug("Endpoint : " + endpointName + " removed from the configuration");
+                if (!cAppArtifactDataService.isArtifactDeployedFromCApp(getTenantId(), ServiceBusConstants.ENDPOINT_TYPE + File.separator + endpointName)) {
+                    if (log.isDebugEnabled()) {
+                        log.debug("Deleting endpoint : " + endpointName + " from the configuration");
+                    }
+                    SynapseConfiguration synapseConfiguration = getSynapseConfiguration();
+                    Endpoint endpoint = synapseConfiguration.getDefinedEndpoints().get(endpointName);
+                    synapseConfiguration.removeEndpoint(endpointName);
+                    MediationPersistenceManager pm = getMediationPersistenceManager();
+                    String fileName = null;
+                    if (endpoint instanceof AbstractEndpoint) {
+                        fileName = endpoint.getFileName();
+                    }
+
+                    pm.deleteItem(endpointName, fileName, ServiceBusConstants.ITEM_TYPE_ENDPOINT);
+
+                    if (log.isDebugEnabled()) {
+                        log.debug("Endpoint : " + endpointName + " removed from the configuration");
+                    }
                 }
             }
 
@@ -254,7 +270,7 @@ public class EndpointAdmin extends AbstractServiceBusAdmin {
      * Delete
      * @throws EndpointAdminException
      */
-     public void deleteAllEndpointGroups() throws EndpointAdminException {
+    public void deleteAllEndpointGroups() throws EndpointAdminException {
         final Lock lock = getLock();
         try {
             lock.lock();
@@ -262,22 +278,28 @@ public class EndpointAdmin extends AbstractServiceBusAdmin {
             Map<String, Endpoint> namedEndpointMap = synapseConfiguration.getDefinedEndpoints();
             Collection<String> namedEndpointCollection = namedEndpointMap.keySet();
             if (namedEndpointCollection.size()>0) {
+                CAppArtifactDataService cAppArtifactDataService = ConfigHolder.getInstance().
+                        getcAppArtifactDataService();
                 for (String endpointName : namedEndpointCollection) {
                     assertNameNotEmpty(endpointName);
                     endpointName = endpointName.trim();
-                    if (log.isDebugEnabled()) {
-                        log.debug("Deleting endpoint : " + endpointName + " from the configuration");
-                    }
-                    Endpoint endpoint = synapseConfiguration.getDefinedEndpoints().get(endpointName);
-                    synapseConfiguration.removeEndpoint(endpointName);
-                    MediationPersistenceManager pm = getMediationPersistenceManager();
-                    String fileName = null;
-                    if (endpoint instanceof AbstractEndpoint) {
-                        fileName = endpoint.getFileName();
-                    }
-                    pm.deleteItem(endpointName, fileName, ServiceBusConstants.ITEM_TYPE_ENDPOINT);
-                    if (log.isDebugEnabled()) {
-                        log.debug("Endpoint : " + endpointName + " removed from the configuration");
+                    if (!cAppArtifactDataService.isArtifactDeployedFromCApp(getTenantId(), ServiceBusConstants.ENDPOINT_TYPE + File.separator + endpointName)) {
+                        if (log.isDebugEnabled()) {
+                            log.debug("Deleting endpoint : " + endpointName + " from the configuration");
+                        }
+                        Endpoint endpoint = synapseConfiguration.getDefinedEndpoints().get(endpointName);
+                        synapseConfiguration.removeEndpoint(endpointName);
+                        MediationPersistenceManager pm = getMediationPersistenceManager();
+                        String fileName = null;
+                        if (endpoint instanceof AbstractEndpoint) {
+                            fileName = endpoint.getFileName();
+                        }
+
+                        pm.deleteItem(endpointName, fileName, ServiceBusConstants.ITEM_TYPE_ENDPOINT);
+
+                        if (log.isDebugEnabled()) {
+                            log.debug("Endpoint : " + endpointName + " removed from the configuration");
+                        }
                     }
                 }
             }
@@ -363,12 +385,21 @@ public class EndpointAdmin extends AbstractServiceBusAdmin {
                 }
             });
 
+            CAppArtifactDataService cAppArtifactDataService = ConfigHolder.getInstance().
+                    getcAppArtifactDataService();
+
             List<EndpointMetaData> metaDatas = new ArrayList<EndpointMetaData>();
             for (Endpoint ep : epList) {
                 EndpointMetaData data = new EndpointMetaData();
                 data.setName(ep.getName());
                 data.setDescription(ep.getDescription());
 
+                if (cAppArtifactDataService.isArtifactDeployedFromCApp(getTenantId(), ServiceBusConstants.ENDPOINT_TYPE + File.separator + ep.getName())) {
+                    data.setDeployedFromCApp(true);
+                }
+                if (cAppArtifactDataService.isArtifactEdited(getTenantId(), ServiceBusConstants.ENDPOINT_TYPE + File.separator + ep.getName())) {
+                    data.setEdited(true);
+                }
                 // Statistics
                 EndpointDefinition def = ((AbstractEndpoint) ep).getDefinition();
                 if (null != def) {
@@ -410,7 +441,11 @@ public class EndpointAdmin extends AbstractServiceBusAdmin {
 
             ((AbstractEndpoint) ep).getDefinition().enableStatistics();
 
-            persistEndpoint(ep);
+            CAppArtifactDataService cAppArtifactDataService = ConfigHolder.getInstance().
+                    getcAppArtifactDataService();
+            if (!cAppArtifactDataService.isArtifactDeployedFromCApp(getTenantId(), ServiceBusConstants.ENDPOINT_TYPE + File.separator + endpointName)) {
+                persistEndpoint(ep);
+            }
             if (log.isDebugEnabled()) {
                 log.debug("Statistics enabled on endpoint : " + endpointName);
             }
@@ -435,8 +470,11 @@ public class EndpointAdmin extends AbstractServiceBusAdmin {
             endpointName = endpointName.trim();
             Endpoint ep = getSynapseConfiguration().getEndpoint(endpointName);
             ((AbstractEndpoint) ep).getDefinition().disableStatistics();
-
-            persistEndpoint(ep);
+            CAppArtifactDataService cAppArtifactDataService = ConfigHolder.getInstance().
+                    getcAppArtifactDataService();
+            if (!cAppArtifactDataService.isArtifactDeployedFromCApp(getTenantId(), ServiceBusConstants.ENDPOINT_TYPE + File.separator + endpointName)) {
+                persistEndpoint(ep);
+            }
             if (log.isDebugEnabled()) {
                 log.debug("Statistics disabled on endpoint : " + endpointName);
             }
@@ -487,6 +525,10 @@ public class EndpointAdmin extends AbstractServiceBusAdmin {
             } catch (XMLStreamException e) {
                 return false;
             }
+
+            CAppArtifactDataService cAppArtifactDataService = ConfigHolder.getInstance().
+                    getcAppArtifactDataService();
+
             if (endpointElement.getQName().getLocalPart()
                     .equals(XMLConfigConstants.ENDPOINT_ELT.getLocalPart())) {
 
@@ -539,7 +581,13 @@ public class EndpointAdmin extends AbstractServiceBusAdmin {
                 endpointName = endpointName.trim();
                 getSynapseConfiguration().removeEndpoint(endpointName);
                 getSynapseConfiguration().addEndpoint(endpointName, endpoint);
-                persistEndpoint(endpoint);
+
+                if (cAppArtifactDataService.isArtifactDeployedFromCApp(getTenantId(), ServiceBusConstants.ENDPOINT_TYPE + File.separator + previousEndpoint.getName())) {
+                    cAppArtifactDataService.setEdited(getTenantId(), ServiceBusConstants.ENDPOINT_TYPE + File.separator + previousEndpoint.getName());
+                } else {
+                    persistEndpoint(endpoint);
+                }
+
                 if (log.isDebugEnabled()) {
                     log.debug("Updated the definition of the endpoint : " + endpointName);
                 }
@@ -549,6 +597,8 @@ public class EndpointAdmin extends AbstractServiceBusAdmin {
             }
         } catch (SynapseException syne) {
             handleFault("Unable to edit Endpoint ", syne);
+        } catch (Exception e) {
+            e.printStackTrace();
         } finally {
             lock.unlock();
         }
@@ -683,7 +733,7 @@ public class EndpointAdmin extends AbstractServiceBusAdmin {
 					return null;
 				}
 				if (obj instanceof OMElement) {
-					e = (OMElement) obj;			
+					e = (OMElement) obj;
 				} else {
 					log.error("Invalid endpoint configuration");
 				}
@@ -774,7 +824,7 @@ public class EndpointAdmin extends AbstractServiceBusAdmin {
             if (registry != null) {
                 if (registry.getRegistryEntry(key).getType() == null) {
                     handleFault("The key '" + key +
-                                "' cannot be found within the configuration", null);
+                            "' cannot be found within the configuration", null);
                 }
                 registry.delete(key);
